kafka-commons/src/main/java/uk/gov/homeoffice/digital/sas/kafka/serializer/CustomJsonSerializer.java
package uk.gov.homeoffice.digital.sas.kafka.serializer;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.SerializationFeature;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import org.springframework.kafka.support.serializer.JsonSerializer;

public class CustomJsonSerializer<T> extends JsonSerializer<T> {

  public static final ObjectMapper mapper = createMapper();

  public static final ObjectMapper createMapper() {
    ObjectMapper mapper = new ObjectMapper();
    mapper.registerModule(new JavaTimeModule());
    mapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
    return mapper;

  }

  public CustomJsonSerializer() {
    super(mapper);
  }

}
